Temporarily always query for the tag
The idea here is - insert the tag - get the inserted row id - use it. If
the tag exists - no inserted row id is returned - SELECT query is
issued.

However the lastInsertId() is always valid even though the INSERT failed
because the tag already exists. Turns out sqlite auto-increments the id
anyway and that seems what is being returned here.

So for now always query for the tag instead.